% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_db.R
\name{build_clique_db}
\alias{build_clique_db}
\title{Create a sqlite database containing all maximal cliques from a network}
\usage{
build_clique_db(ppi_network, db_folder, db_name)
}
\arguments{
\item{ppi_network}{A network as a dataframe where the first 2 columns are the interactions}

\item{db_folder}{A directory where the database will be stored. 
Please not that tilde expension does not work.}

\item{db_name}{The name of the database. File suffix ".sqlite" will be appended}
}
\description{
Create a sqlite database containing all maximal cliques from a network
}
\details{
Creates a SQLite database containing all maximal cliques from the network in the folder 
\code{db_folder} with filename \code{db_name.sqlite}. This database can be used as in 
input to \code{\link{clique_sum_exact}} and \code{\link{clique_sum_permutation}}
}
\author{
Dirk de Weerd
}
