% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_sum.R
\name{clique_sum}
\alias{clique_sum}
\title{Clique Sum}
\usage{
clique_sum(MODifieR_input, ppi_network, simplify_graph = T, n_iter = 10000,
  cutoff = 0.05, min_clique_size = 5, min_deg_clique = 3,
  dataset_name = NULL)
}
\arguments{
\item{n_iter}{number of iterations to be performed}

\item{cutoff}{Desired P-value cutoff}

\item{min_clique_size}{Numeric shortest clique size (default = 5)}

\item{min_deg_clique}{Numeric Minimum degree of the clique}

\item{deg_genes}{A nx2 dataframe consisting of differentially expressed genes and their respective -log10p values}

\item{igraph_network}{A path to dataframe csv PPi network of your choice}
}
\value{
A MODifieR class object with disease module and settings
}
\description{
A Clique based Algorithm proposed by Fredrik Barranais to produce a disease module from Differentially expressed genes using a core sum approach
}
