% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_sum_permutation.R
\name{clique_sum_permutation}
\alias{clique_sum_permutation}
\title{Clique Sum}
\usage{
clique_sum_permutation(MODifieR_input, db, n_iterations = 10000,
  clique_significance = 0.01, min_clique_size = 5,
  multiple_cores = T, n_cores = 4, dataset_name = NULL)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by one of the \code{create_input} functions}

\item{db}{A clique database created by \code{\link{build_clique_db}}}

\item{n_iterations}{Number of iterations to be performed for the permutation based p-value}

\item{clique_significance}{p-value for cliques to be considered significant}

\item{min_clique_size}{Minimal size for cliques}

\item{multiple_cores}{Parallel process using multiple cores?}

\item{n_cores}{Number of cores to use if parallel processing}

\item{dataset_name}{Optional name for the input object that will be stored in the settings object.
Default is the variable name of the input object}
}
\value{
clique_sum_permutation returns an object of class "MODifieR_module" 
with subclass "Clique_Sum_permutation". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
An implementation of the clique-based disease module inference method proposed by Gustafsson et al.
}
\details{
Clique_sum_permutation finds cliques of at least size \code{min_clique_size} that are 
 significantly enriched with DEGs. For every clique size, a null distribution is created using
 the summed -log 10 p-values. The union of maximal cliques with a summed -log 10 p-value below \code{clique_significance} and at 
 least \code{min_deg_in_clique} is the final disease module.
}
\references{
\cite{Gustafsson, M., Edström, M., Gawel, D., Nestor, C. E., Wang, H., Zhang, H., … Benson, M. (2014). 
Integrated genomic and prospective clinical studies show the importance of modular pleiotropy for disease susceptibility, 
diagnosis and treatment. Genome Medicine, 6(2), 17. https://doi.org/10.1186/gm534}
}
\author{
Dirk de Weerd
}
