% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_clique.R
\name{correlation_adjust_cutoff}
\alias{correlation_adjust_cutoff}
\title{correlation_adjust_cutoff}
\usage{
correlation_adjust_cutoff(frequency_cutoff, correlation_module)
}
\arguments{
\item{frequency_cutoff}{significance out of number of iterations performed (default = 0.05)}

\item{wgcna_module}{Module object that has been produced by \code{correlation_clique}
function}
}
\value{
\code{correlation_clique} module object
}
\description{
correlation_adjust_cutoff
}
\details{
This function allows to adjust the frequency cutoff for a \code{correlation_clique} module object
}
\seealso{
\code{\link{correlation_clique}}
}
