% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_clique.R
\name{correlation_clique}
\alias{correlation_clique}
\title{Clique_correlation}
\usage{
correlation_clique(MODifieR_input, ppi_network, frequency_cutoff = 0.5,
  cutOffRank = 700, probabilityScaleFactor = 0.6, iteration = 50,
  signif_cutoff = 0.01, gene_significance = 0.05, dataset_name = NULL)
}
\arguments{
\item{ppi_network}{A dataframe of PPI network of your choice}

\item{frequency_cutoff}{Numeric significance out of number of iterations performed (default = 0.05)}

\item{iteration}{Number of iterations to be performed}

\item{signif_cutoff}{Cutoff for Fisher exact test for cliques}

\item{gene_significance}{Cutoff for significant genes}

\item{pval_matrix}{A nx2 dataframe of the entrez ids and their respective p-values}

\item{correlation_matrix}{A nxn matrix of Correlated p-values of n genes}

\item{cutoffRank}{Score cutoff of the string PPI network}

\item{cliquetestCriteria}{type of test to be perfomed for enrichment of cliques (default = "FisherTest")}

\item{probabailitySacleFactor}{Scale for enriched cliques}
}
\value{
A MODifieR class object with disease module and settings
}
\description{
A clique based method to find a disease module from correlated gene expression
}
