% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_clique.R
\name{correlation_set_module_size}
\alias{correlation_set_module_size}
\title{correlation_adjust_cutoff}
\usage{
correlation_set_module_size(size, correlation_module)
}
\arguments{
\item{size}{Module object that has been produced by \code{correlation_clique}
function}
}
\value{
\code{correlation_clique} module object
}
\description{
correlation_adjust_cutoff
}
\details{
The function will find the the frequency cutoff for that will result
in a \code{correlation_clique} module object closest to \code{size}
}
\seealso{
\code{\link{correlation_clique}}
}
