% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_matrix}
\alias{count_matrix}
\title{Count matrix}
\format{An object of class \code{matrix} with 16569 rows and 10 columns.}
\usage{
count_matrix
}
\description{
An example RNA-seq dataset from GEO (accession number GSE123496) containing
Multiple Sclerosis (MS) patients and healthy controls
}
\details{
The count matrix contains 10 RNA-seq samples. Columns 1:5 contain RNA-seq 
data from the frontal cortex of MS patients, columns 6:10 are matched controls. The dataset
has been filtered, all genes with less than 100 counts per row are removed.
There are 16569 genes in the dataset.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE123496}.
}
\seealso{
\code{\link{ppi_network}}
}
\author{
Dirk de Weerd
}
\keyword{datasets}
