% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input_rnaseq}
\alias{create_input_rnaseq}
\title{Creates an RNA-seq input object for downstream analysis}
\usage{
create_input_rnaseq(count_matrix, group1_indici, group2_indici,
  group1_label, group2_label, expression = T,
  differential_expression = T, use_adjusted = T,
  normalize_quantiles = F)
}
\arguments{
\item{count_matrix}{Matrix containing raw RNA-seq counts}

\item{group1_indici}{vector containing indici for samples belonging to group 1 (Column numbers)}

\item{group2_indici}{vector containing indici for samples belonging to group 2 (Column numbers)}

\item{group1_label}{Label for each group 1, for example "patient" or "control"}

\item{group2_label}{Label for each group 2, for example "patient" or "control"}

\item{expression}{boolean, calculate expression values?}

\item{differential_expression}{boolean, calculate differentially expressed data?}

\item{use_adjusted}{boolean, use adjusted p value for differential expression analysis?}

\item{normalize_quantiles}{boolean, Normalize quantiles for WGCNA-based methods?}
}
\value{
The function returns an object of class "MODifieR_input". The object is a named list containing the
following components:
\item{diff_genes}{A 2 two column data.frame where the first column are genes and the second column unadjusted p-values
obtained by differential expression analysis}
\item{edgeR_deg_table}{A data.frame from \code{edgeR glmQLFit}}
\item{annotated_exprs_matrix}{A matrix where the rows are genes and the columns samples. Raw counts have been normalized
with \code{varianceStabilizingTransformation} and optionally quantile normalized}
\item{count_matrix}{A matrix, the original input count matrix}
\item{group_indici}{A named list containing 2 numeric vectors. The names are the group labels and the values 
are the group indici}
}
\description{
The MODifieR input objects can be used in downstream analysis for the disease module inference methods included
in this package.
}
\details{
The function creates an input object to be used in all disease module inference methods. Differentially
expressed genes are calculated using generalized linear models from the \code{edgeR} package. For WGCNA-based
methods raw counts are normalized using the \code{varianceStabilizingTransformation} from the \code{DESeq2} package. 
Optionally, Quantile normalization using the \code{normalize.quantiles} function from the \code{preprocessCore}
can be applied.
}
\seealso{
\code{\link[edgeR]{glmQLFit}}
\code{\link[DESeq2]{varianceStabilizingTransformation}}
}
