% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{diamond}
\alias{diamond}
\title{DIAMOnD module}
\usage{
diamond(MODifieR_input, ppi_network, deg_cutoff = 0.05,
  n_output_genes = 200, seed_weight = 10, include_seed = T,
  tempfile_genes = tempfile(), dataset_name = NULL)
}
\arguments{
\item{ppi_network}{path to the desired background ppi network}

\item{deg_cutoff}{p value cutoff for differentially expressed genes}

\item{n_output_genes}{maximum number of genes to be included in the final module}

\item{seed_weight}{Numeric additional parameter to assign weight for the seed genes}

\item{include_seed}{Logical TRUE/FALSE for inclusion of seed genes in the output module}

\item{input_genes}{path to the tab delimited file consisting of seed genes}
}
\value{
A disease module
}
\description{
A seed gene based algorithm to identify disease module from Differentially Expressed Genes
}
