% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcoex.R
\name{diffcoex_split_module_by_color}
\alias{diffcoex_split_module_by_color}
\title{Returns new DiffCoEx module objects by color}
\usage{
diffcoex_split_module_by_color(diffcoex_module)
}
\arguments{
\item{diffcoex_module}{Module object that has been produced by \code{diffcoex} function}
}
\value{
A list of \code{DiffCoEx} module objects
}
\description{
Returns new DiffCoEx module objects by color
}
\details{
The  \code{DiffCoEx} module object is split into a series of \code{DiffCoEx} objects by color.
Eevery significant color in the module will be its own \code{DiffCoEx} module object
}
\seealso{
\code{\link{diffcoex}}
}
