% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{entrez_to_symbol}
\alias{entrez_to_symbol}
\title{Convert the module genes in a MODifieR_input object from ENTREZ gene IDs to official gene symbols}
\usage{
entrez_to_symbol(MODifieR_module)
}
\arguments{
\item{MODifieR_module}{An object of class MODifieR_module}
}
\description{
Convert the module genes in a MODifieR_input object from ENTREZ gene IDs to official gene symbols
}
\details{
The function uses the \code{org.Hs.egSYMBOL} function from the package \code{org.Hs.eg.db}
to convert official gene symbols to ENTREZ IDs
}
\seealso{
\code{\link[org.Hs.eg.db]{org.Hs.egSYMBOL}}

\code{\link{symbol_to_entrez}}
}
