% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_matrix}
\alias{expression_matrix}
\title{Expression matrix}
\format{An object of class \code{matrix} with 52307 rows and 16 columns.}
\usage{
expression_matrix
}
\description{
An example micorarray dataset from GEO (accession number GSE4588) containing
systemic lupus erythematosus patients and healthy controls
}
\details{
The expression matrix contains 16 microarray samples. Columns 1:9 are healthy controls,
columns 10:16 are SLE patients. There are 52307 probes in this dataset.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4588}.
}
\seealso{
\code{\link{probe_annotation}}

\code{\link{ppi_network}}
}
\author{
Dirk de Weerd
}
\keyword{datasets}
