% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_set.R
\name{get_max_frequency}
\alias{get_max_frequency}
\title{Get maximal number of MODifieR modules that share at least one gene}
\usage{
get_max_frequency(module_list)
}
\arguments{
\item{module_list}{A list of MODifieR modules}
}
\value{
The function returns an integer that gives the maximal number of MODifieR objects in 
\code{module_list} that share at least one gene.
}
\description{
Get maximal number of MODifieR modules that share at least one gene
}
