% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcode.R
\name{mod_mcode}
\alias{mod_mcode}
\title{MCODE}
\usage{
mod_mcode(MODifieR_input, ppi_network, type = "Gene symbol", hierarchy = 1,
  vwp = 0.5, haircut = F, fluff = F, fdt = 0.8, loops = T,
  diffgen_cutoff = 1.3, module_cutoff = 3.5, dataset_name = NULL)
}
\arguments{
\item{type}{keytype of the gene that is provide (default = "Gene Symbol")}

\item{hierarchy}{the hierarchical level of the PPI network to be used (default = 1)}

\item{vwp}{Vertex weight percentage. Default value is 0.5.}

\item{haircut}{Boolean value, whether to remove singly-connected nodes from clusters (TRUE) or not (FALSE).}

\item{fluff}{Boolean value, whether to spand cluster cores by one neighbour shell outwards (TRUE) or not (FALSE).}

\item{fdt}{Cluster density cutoff. Default value is 0.8.}

\item{loops}{Boolean value, whether to include self-loops (TRUE) or not (FALSE).}

\item{deg_genes}{A nx2 dataframe sorted ranked list of genes with their respective -log10 p-values}
}
\description{
A clique based Algorithm to identify disease module from differentially expressed genes
}
