% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{MODA}
\usage{
moda(MODifieR_input, CuttingCriterion = "Density", specificTheta = 0.1,
  conservedTheta = 0.1, result_folder, dataset_name = NULL)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by \code{\link{create_input}} function}

\item{datExpr1}{A dataframe of the microarray matrix of the Control samples}

\item{datExpr2}{A dataframe of the microarray matrix of the Control samples}
}
\description{
A Coexpression based algorithm to identify disease module
}
