% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda_change_specific_threshold}
\alias{moda_change_specific_threshold}
\title{Change}
\usage{
moda_change_specific_threshold(moda_module, specificTheta)
}
\arguments{
\item{moda_module}{A \code{MODifieR_input} object created by \code{\link{moda}}}

\item{specificTheta}{the threshold to define min(s)+specificTheta, 
less than which is considered as condition specific module. 
s is the sums of rows in Jaccard index matrix. See supplementary file.}
}
\description{
Change
}
\author{
Dirk de Weerd
}
