% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaal.R
\name{modaal}
\alias{modaal}
\title{MODA}
\usage{
modaal(MODifieR_input, CuttingCriterion = "Density",
  specificTheta = 0.1, conservedTheta = 0.1, dataset_name = NULL,
  group_of_interest)
}
\arguments{
\item{datExpr1}{A dataframe of the microarray matrix of the Control samples}

\item{datExpr2}{A dataframe of the microarray matrix of the Control samples}
}
\description{
A Coexpression based algorithm to identify disease module
}
