% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_network}
\alias{ppi_network}
\title{PPi network}
\format{An object of class \code{data.frame} with 64672 rows and 3 columns.}
\usage{
ppi_network
}
\description{
A Protein-Protein interaction network from STRING
}
\details{
This PPi network from STRING is version 7.1 and filtered for interactions with a 
confidence score higher than 700. The identifiers for the genes are ENTREZ.
There are 64672 interactions (rows) in the dataframe. Each row denotes an interaction;
column 1 contains the first gene, column 2 the second. The third column is the confidence score.
}
\references{
\url{https://string-db.org/}
}
\seealso{
\code{\link{probe_annotation}}

\code{\link{expression_matrix}}
}
\author{
Dirk de Weerd
}
\keyword{datasets}
