% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_annotation}
\alias{probe_annotation}
\title{Probe annotation}
\format{An object of class \code{data.frame} with 52307 rows and 2 columns.}
\usage{
probe_annotation
}
\description{
Probe annotation for the example dataset \code{\link{expression_matrix}}.
The annotation is taken from Bioconductor package hgu133plus2.db and contains 52307 probes.
}
\references{
\url{https://bioconductor.org/packages/release/data/annotation/html/hgu133plus2.db.html}

\cite{Carlson M (2016). hgu133plus2.db: Affymetrix Human Genome U133 Plus 2.0 
Array annotation data (chip hgu133plus2). R package version 3.2.3.}
}
\seealso{
\code{\link{expression_matrix}}

\code{\link{ppi_network}}
}
\keyword{datasets}
