% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{recalculate_diff_genes}
\alias{recalculate_diff_genes}
\title{Recalculate DEGs}
\usage{
recalculate_diff_genes(MODifieR_input, use_adjusted)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by \code{\link{create_input}} function}

\item{use_adjusted}{boolean, use adjusted p value for differential expression analysis?}
}
\value{
MODifieR_input object
}
\description{
Recalculate DEGs
}
\details{
Recalculate DEGs to either use adjusted or unadjusted p values
}
\seealso{
\code{\link{create_input}}
}
