% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{recalculate_expression}
\alias{recalculate_expression}
\title{Recalculate collapsing probes to genes}
\usage{
recalculate_expression(MODifieR_input, method)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by one of the \code{create_input} functions}

\item{method}{character string for determining which method is used to choose a probe among
exactly 2 corresponding rows or when connectivityBasedCollapsing=FALSE. These are the options:
"MaxMean" (default) or "MinMean" = choose the row with the highest or lowest mean value, respectively.
"maxRowVariance" = choose the row with the highest variance (across the columns of \code{datET}). 
"absMaxMean" or "absMinMean" = choose the row with the highest or lowest mean absolute value.
"ME" = choose the eigenrow (first principal component of the rows in each group).  Note that with this
       method option, connectivityBasedCollapsing is automatically set to FALSE.
"Average" = for each column, take the average value of the rows in each group 
"function" = use this method for a user-input function (see the description of the argument
"methodFunction").
Note: if method="ME", "Average" or "function", the output parameters "group2row" and "selectedRow" are not informative.
}
}
\value{
MODifieR_input object
}
\description{
Recalculate collapsing probes to genes
}
\details{
Recalculate the collapsing of probes to genes using on the \code{method} options
}
\seealso{
\code{\link{create_input}}
}
\author{
Dirk de Weerd
}
