% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{remove_diamond_seed_genes}
\alias{remove_diamond_seed_genes}
\title{Remove seed genes from a DIAMOnD MODifieR_module}
\usage{
remove_diamond_seed_genes(diamond_module)
}
\arguments{
\item{diamond_module}{A \code{MODifieR_input} object created by \code{\link{diamond}}}
}
\value{
An object of class "MODifieR_module" with subclass "DIAMOnD"
}
\description{
Remove seed genes from a DIAMOnD MODifieR_module
}
\details{
Removes seed genes from a DIAMOnD module
}
\seealso{
\code{\link{diamond}}
}
