% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2B.R
\name{S2B}
\alias{S2B}
\title{S2B}
\usage{
S2B(MODifieR_module1, MODifieR_module2, ppi_network, nrep, nrep2, n_cores)
}
\arguments{
\item{MODifieR_module1}{A MODifieR module inferred by a disease module inference method}

\item{MODifieR_module2}{A MODifieR module inferred by a disease module inference method}

\item{ppi_network}{A network as a dataframe where the first 2 columns are the interactions}

\item{nrep}{number of randomizations (shuffle edges maintaining node degree) 
to compute specificity score 1}

\item{nrep2}{number of randomizations (shuffle seed identity) to compute specificity score 2}

\item{n_cores}{Number of cores to use}
}
\value{
Returns an object of class "MODifieR_module" with subclass "S2B". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{input1_class}{Method that was used to create disease module 1 (inferred by its class)}
\item{input2_class}{Method that was used to create disease module 2 (inferred by its class)}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
Combine two different disease modules into using Specific Betweenness (S2B)
}
\details{
S2B prioritizes genes frequently and specifically present in shortest paths linking two disease modules. 
For a detailed description of the algorithm, see the paper referenced below
}
\references{
\cite{Garcia-Vaquero, M. L., Gama-Carvalho, M., Rivas, J. D. Las, & Pinto, F. R. (2018). 
Searching the overlap between network modules with specific betweeness (S2B) and its application 
to cross-disease analysis. Scientific Reports, 8(1), 1–10. 
\url{https://doi.org/10.1038/s41598-018-29990-7}}
}
