% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_module.R
\name{super_module}
\alias{super_module}
\title{Robust module concensus}
\usage{
super_module(module_list, include_single_methods = T)
}
\arguments{
\item{module_list}{A list object with f´genes of each module}

\item{include_single_methods}{A boolean value , whether to include the original modules in the final result}
}
\description{
A final conglomeration of all modules identified by different module methods to produce all kinds of intersect modules
}
