% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{symbol_to_entrez}
\alias{symbol_to_entrez}
\title{Convert the module genes in a MODifieR_input object from official gene symbols to ENTREZ gene IDs}
\usage{
symbol_to_entrez(MODifieR_module)
}
\arguments{
\item{MODifieR_module}{An object of class MODifieR_module}
}
\description{
Convert the module genes in a MODifieR_input object from official gene symbols to ENTREZ gene IDs
}
\details{
The function uses the \code{org.Hs.egSYMBOL} function from the package \code{org.Hs.eg.db}
to convert ENTREZ IDs to official gene symbols
}
\seealso{
\code{\link[org.Hs.eg.db]{org.Hs.egSYMBOL}}

 \code{\link{entrez_to_symbol}}
}
