% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{wgcna}
\alias{wgcna}
\title{WGCNA}
\usage{
wgcna(MODifieR_input, pval_cutoff = 0.05, min_module_size = 30,
  deep_split = 2, pam_respects_dendro = F, merge_cut_height = 0.1,
  numeric_labels = T, min_KME = 0, save_toms = T,
  dataset_name = deparse(substitute(MODifieR_input)))
}
\arguments{
\item{dataset1}{A dataframe of the microarray matrix of the Control samples}

\item{dataset2}{A dataframe of the microarray matrix of the Patient samples}

\item{traits_file}{A dataframe of the associted traits/phenotypes  for the included samples}
}
\description{
A Coexpression based Algorithm to identify disease module
}
