% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{wgcna_adjust_significance}
\alias{wgcna_adjust_significance}
\title{wgcna_adjust_significance}
\usage{
wgcna_adjust_significance(pval_cutoff, wgcna_module, use_unadjusted = F)
}
\arguments{
\item{pval_cutoff}{The p-value cutoff to be used for significant co-expression modules (colors)}

\item{wgcna_module}{Module object that has been produced by \code{wgcna} function}

\item{use_unadjusted}{Boolean value to signify if the adjusted (TRUE) or unadjusted (FALSE)
p value should be used to adjust significance}
}
\value{
\code{wgcna} module object
}
\description{
wgcna_adjust_significance
}
\details{
This function allows to adjust the significance cutoff for a \code{wgcna} module object
}
\seealso{
\code{\link{wgcna}}
}
