% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{wgcna_get_all_module_genes}
\alias{wgcna_get_all_module_genes}
\title{Generate a list of module colors with their respective genes.}
\usage{
wgcna_get_all_module_genes(wgcna_module)
}
\arguments{
\item{wgcna_module}{Module object that has been produced by \code{wgcna} function}
}
\value{
Returns a \code{named list} of module genes where the names are the module colors. Includes 
non-significant colors
}
\description{
Generate a list of module colors with their respective genes.
}
\seealso{
\code{\link{wgcna}}
}
\author{
Dirk de Weerd
}
