% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna_trait.R
\name{wgcna_get_module_genes_by_sign}
\alias{wgcna_get_module_genes_by_sign}
\title{Split WGCNA module in module containing only positive or negative correlation}
\usage{
wgcna_get_module_genes_by_sign(wgcna_module, mode)
}
\arguments{
\item{wgcna_module}{Module object that has been produced by \code{wgcna_trait} function}

\item{mode}{Character. "p" or "positive" for positive correlation, "n" or "negative" 
for negative correlation.}
}
\value{
\code{wgcna_module} object
}
\description{
Split WGCNA module in module containing only positive or negative correlation
}
\details{
The functions returns a new \code{wgcna} module object that only contains positively or 
negatively correlated colors
}
\seealso{
\code{\link{wgcna_trait}}
}
