% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna_trait.R
\name{wgcna_set_module_size}
\alias{wgcna_set_module_size}
\title{Returns a wgcna module close to \code{parameter}}
\usage{
wgcna_set_module_size(size, wgcna_module)
}
\description{
Returns a wgcna module close to \code{parameter}
}
