% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{wgcna_split_module_by_color}
\alias{wgcna_split_module_by_color}
\title{Returns new WGCNA module objects by color}
\usage{
wgcna_split_module_by_color(wgcna_module)
}
\arguments{
\item{wgcna_module}{Module object that has been produced by \code{wgcna} function}
}
\value{
A list of \code{wgcna} module objects
}
\description{
Returns new WGCNA module objects by color
}
\details{
The  \code{wgcna} module object is split into a series of \code{wgcna} objects by color.
Eevery significant color in the module will be its own \code{wgcna} module object
}
\seealso{
\code{\link{wgcna}}
}
