% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_clique.R
\name{correlation_clique}
\alias{correlation_clique}
\title{Clique_correlation}
\usage{
correlation_clique(MODifieR_input, ppi_network, frequency_cutoff = 0.5,
  fraction_of_interactions = 0.4, iteration = 50,
  clique_significance = 0.01, deg_cutoff = 0.05, multiple_cores = F,
  n_cores = 3, dataset_name = NULL)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by one of the \code{create_input} functions}

\item{ppi_network}{A network as a dataframe where the first 2 columns are the interactions}

\item{frequency_cutoff}{Fraction of the number of times a gene should be present in it 
iterations. Default is 0.5, meaning 50 procent of all iterations}

\item{fraction_of_interactions}{Fraction of interactions from the original network that will be used
in each iteration}

\item{iteration}{Number of iterations to be performed}

\item{clique_significance}{Cutoff for Fisher exact test for cliques}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{multiple_cores}{parallelize iterations using number of cores on system -1?}

\item{n_cores}{Number of cores to use}

\item{dataset_name}{Optional name for the input object that will be stored in the settings object.
Default is the variable name of the input object}

\item{tempfolder}{Folder where temporary files are stored}
}
\value{
correlation_clique returns an object of class "MODifieR_module" with subclass "Correlation_clique". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{frequency_table}{A table containing the fraction of times the genes were present in an iteration module}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
A clique based method to find a disease module from correlated gene expression
}
\details{
The correlation clique is a clique-based algorithm using consensus clustering.
The algorithm starts with calculating a correlation score between each interaction in the PPi network.
The correlation score is obtained by subtracting the Pearson correlation p-value:
\deqn{correlation score = 1 - Pearson p-value}
Subsequently, the correlation score is multiplied by 
the correlation confidence and scaled with the scale factor to get the edge score:
\deqn{edge score = \sqrt(correlation score * confidence score  ) * scale_factor}
When the edge scores are calculated the iterative part of the algorithm commences:
All edge scores are compared to random variables from the uniform distribution between (0,1)
only interactions where the edge score is higher than the random variable are used to construct
a new PPi network. Then, maximal cliques are inferred from this new network.
The cliques are tested for significant enrichment of DEGs by Fisher's exact test and
the union of significant cliques is the disease module for this iteration
The final disease module will consist of genes that have been present in at least \code{frequency_cutoff} iterations
}
\author{
Dirk de Weerd
}
