% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input}
\alias{create_input}
\title{Creates an object of class "MODifieR input"}
\usage{
create_input(expression_matrix, probe_map, group1_indici, group2_indici,
  group1_label, group2_label, expression = F, diff_data = F,
  correlation_clique = F)
}
\arguments{
\item{expression_matrix}{Normalized expression matrix where the samples are columns and probes are rows}

\item{probe_map}{A dataframe providing annotation for the probes. The dataframe should have 3 columns: 
\itemize{
\item {PROBEID}: The probe id as it is in the expression matrix
\item {SYMBOL}: The gene symbol (if available) associated with the probe
\item{ENTREZID}: The entrez id (if available) associated with the probe
}}

\item{expression}{boolean, calculate expression values?}

\item{diff_data}{boolean, calculate differentially expressed data?}

\item{correlation_clique}{boolean, calculate correlation matrix?}

\item{group_indici}{vectors containing indici for different groups (Column numbers)}

\item{group_labels}{Labels for each group, for example "patient" and "control"}
}
\description{
The MODifieR input object can be used
in downstream analysis using the disease module inference methods included
in the MODifieR package.
}
