% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{diamond}
\alias{diamond}
\title{DIAMOnD module}
\usage{
diamond(MODifieR_input, ppi_network, deg_cutoff = 0.05,
  n_output_genes = 200, seed_weight = 10, include_seed = FALSE,
  dataset_name = NULL)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by \code{\link{create_input}} function}

\item{ppi_network}{A PPi network as a 2 column dataframe}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{n_output_genes}{maximum number of genes to be included in the final module}

\item{seed_weight}{Numeric additional parameter to assign weight for the seed genes}

\item{include_seed}{Logical TRUE/FALSE for inclusion of seed genes in the output module}

\item{dataset_name}{Optional name for the input object that will be stored in the settings object.
Default is the variable name of the input object}
}
\value{
diamond returns an object of class "MODifieR_module" with subclass "DIAMOnD". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{seed_genes}{Character vector containing genes that have been used as seed genes in the algorithm}
\item{ignored_genes}{Potential seed genes that are not in the PPi network}
\item{added_genes}{A table containing information on all added genes. First column is the name of the gene,
the second column is the degree of the node (gene). The third column is the number of +1 neighbors 
and the fourth column is the p-value.}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
A seed gene based algorithm to identify disease module from differentially expressed genes
}
\details{
A slightly modified version of the original DIAMOnD python script is called from within R.
The only change to the orginal algorithm is the option to include the seed genes
to the module. There are also function to add or remove the seed genes from the output object, namely:
\code{\link{diamond_add_seed_genes}} and \code{\link{diamond_remove_seed_genes}}
For a detailed description of how the algorithm works, please see the paper referenced below.
}
\references{
\cite{Ghiassian, S. D., Menche, J., & Barabási, A. L. (2015). 
A DIseAse MOdule Detection (DIAMOnD) Algorithm Derived from a Systematic Analysis of 
Connectivity Patterns of Disease Proteins in the Human Interactome. PLoS Computational 
Biology, 11(4), 1–21. \url{https://doi.org/10.1371/journal.pcbi.1004120}}
}
