% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcoex.R
\name{diffcoex}
\alias{diffcoex}
\title{DiffCoEx}
\usage{
diffcoex(MODifieR_input, value = 20, beta = 6, cor_method = "spearman",
  cluster_method = "average", cuttree_method = "hybrid",
  cut_height = 0.996, deep_split = 0, pam_respects = F,
  min_cluster_size = 20, cluster_merge_cut_height = 0.2,
  dataset_name = NULL)
}
\arguments{
\item{cor_method}{correlation method to be used (default is "spearman")}

\item{dataset1}{A dataframe of the microarray matrix of the Control samples}

\item{dataset2}{A dataframe of the microarray matrix of the Patient samples}
}
\description{
A Coexpression based Algorithm to identify disease module
}
