% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulediscoverer.R
\name{modulediscoverer}
\alias{modulediscoverer}
\title{Module Discoverer}
\usage{
modulediscoverer(MODifieR_input, ppi_network, permutations = 10000,
  deg_cutoff = 0.05, repeats = 15, clique_cutoff = 0.01,
  dataset_name = NULL)
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by one of the \code{create_input} functions}

\item{ppi_network}{A network as a dataframe where the first 2 columns are the interactions}

\item{permutations}{Number of permutations to perform to identify the community structure}

\item{deg_cutoff}{p-value cutoff for differentialy expressed genes}

\item{repeats}{Number of times the algorithm is repeated}

\item{clique_cutoff}{cutoff pvalue for significant cliques}

\item{dataset_name}{Optional name for the input object that will be stored in the settings object.
Default is the variable name of the input object}
}
\value{
modulediscoverer returns an object of class "MODifieR_module" with subclass "module_discoverer". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{graph}{\code{\link[igraph]{igraph}} graph containing the disease module}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
A clique based algorithm by Vlaic et al. to produce disease module from Differentially Expressed Genes
}
\details{
This is an implementation of the \emph{single seed} Module Discoverer algorithm.
The code has been adapted from the orignal code by Vlaic et al. For details, please see the paper referenced below
}
\references{
\cite{Vlaic, S., Tokarski-schnelle, C., Gustafsson, M., Dahmen, U., Guthke, R., & Schuster, S. (2017). 
ModuleDiscoverer: Identification of regulatory modules in protein-protein interaction networks., 1–17.}
}
\seealso{
\url{https://www.leibniz-hki.de/en/modulediscoverer.html}
}
