% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{wgcna}
\alias{wgcna}
\title{An implementation of WGCNA to correlate coexpression modules to disease}
\usage{
wgcna(MODifieR_input, group_of_interest, minModuleSize = 30,
  deepSplit = 2, pamRespectsDendro = F, mergeCutHeight = 0.1,
  numericLabels = T, pval_cutoff = 0.05, corType = "bicor",
  maxBlockSize = 5000, TOMType = "signed", saveTOMs = T,
  maxPOutliers = 0.1,
  dataset_name = deparse(substitute(MODifieR_input)))
}
\arguments{
\item{MODifieR_input}{A MODifieR input object produced by one of the \code{create_input} functions}

\item{group_of_interest}{Numerical value denoting which group contains the condition of interest (1 or 2)}

\item{minModuleSize}{ minimum module size for module detection. See
\code{\link[dynamicTreeCut]{cutreeDynamic}} for more details.  }

\item{deepSplit}{ integer value between 0 and 4. Provides a simplified control over how sensitive
module detection should be to module splitting, with 0 least and 4 most sensitive. See
\code{\link[dynamicTreeCut]{cutreeDynamic}} for more details. }

\item{pamRespectsDendro}{Logical, only used when \code{pamStage} is \code{TRUE}. 
If \code{TRUE}, the PAM stage will
respect the dendrogram in the sense an object can be PAM-assigned only to clusters that lie below it on
the branch that the object is merged into. 
     See \code{\link[dynamicTreeCut]{cutreeDynamic}} for more details. }

\item{mergeCutHeight}{ dendrogram cut height for module merging. }

\item{numericLabels}{ logical: should the returned modules be labeled by colors (\code{FALSE}), or by
numbers (\code{TRUE})? }

\item{pval_cutoff}{The p-value cutoff to be used for significant co-expression modules (colors)}

\item{corType}{character string specifying the correlation to be used. Allowed values are (unique
abbreviations of) \code{"pearson"} and \code{"bicor"}, corresponding to Pearson and bidweight
midcorrelation, respectively. Missing values are handled using the \code{pairwise.complete.obs} option. }

\item{maxBlockSize}{integer giving maximum block size for module detection. Ignored if \code{blocks}
above is non-NULL. Otherwise, if the number of genes in \code{datExpr} exceeds \code{maxBlockSize}, genes
will be pre-clustered into blocks whose size should not exceed \code{maxBlockSize}. }

\item{TOMType}{ one of \code{"none"}, \code{"unsigned"}, \code{"signed"}. If \code{"none"}, adjacency
will be used for clustering. If \code{"unsigned"}, the standard TOM will be used (more generally, TOM
function will receive the adjacency as input). If \code{"signed"}, TOM will keep track of the sign of
correlations between neighbors. }

\item{saveTOMs}{ logical: should the consensus topological overlap matrices for each block be saved
and returned?  }

\item{maxPOutliers}{ only used for \code{corType=="bicor"}. Specifies the maximum percentile of data 
that can be considered outliers on either 
side of the median separately. For each side of the median, if
higher percentile than \code{maxPOutliers} is considered an outlier by the weight function based on
\code{9*mad(x)}, the width of the weight function is increased such that the percentile of outliers on
that side of the median equals \code{maxPOutliers}. Using \code{maxPOutliers=1} will effectively disable
all weight function broadening; using \code{maxPOutliers=0} will give results that are quite similar (but
not equal to) Pearson correlation. }

\item{dataset_name}{Optional name for the input object that will be stored in the settings object.
Default is the variable name of the input object}
}
\value{
wgcna returns an object of class "MODifieR_module" with subclass "WGCNA". 
This object is a named list containing the following components:
\item{module_genes}{A character vector containing the genes in the final module}
\item{info_table}{A data.frame containing all genes and their assigned colors}
\item{correlation_to_trait_table}{A data.frame containing all module colors and their p- and adjusted p-value}
\item{softthreshold_value}{A numeric, the soft threshold power that is used. See: \code{\link[WGCNA]{pickSoftThreshold}}}
\item{module_colors}{A character vector containing the colors that make up the final disease module}
\item{settings}{A named list containing the parameters used in generating the object}
}
\description{
An implementation of WGCNA to correlate coexpression modules to disease
}
\details{
wgcna is an implementation of WGCNA that
associates co-expression modules (denoted by color) to a trait. Co-expression modules with an
adjusted p-value < \code{pval_cutoff} will make up the final disease module.

The algorithm infers co-expression modules from combined expression dataset from both \code{group1}
and \code{group2}.
Co-expression modules are then correlated to trait (group 1 ~ group 2). 

After analysis there are some post-processing functions available:

\itemize{
\item{\code{\link{wgcna_get_all_module_genes}}} Get a list with all genes sorted by module color
\item{\code{\link{wgcna_get_module_genes_by_sign}}} Get a module with either only postively correlated
genes or negatively correlated genes
\item{\code{\link{wgcna_adjust_significance}}} Adjust p-value cutoff
\item{\code{\link{wgcna_split_module_by_color}}} Get a list where each color is a separate module
\item{\code{\link{wgcna_set_module_size}}} Get a module close to a specific size
}
}
\references{
\cite{Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. 
BMC Bioinformatics 2008, 9:559 \url{doi:10.1186/1471-2105-9-559} }


\cite{Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. J
ournal of Statistical Software, 46(11), 1-17. URL \url{http://www.jstatsoft.org/v46/i11/}}
}
\author{
Dirk de Weerd
}
