% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna_trait.R
\name{wgcna_get_module_genes_by_sign}
\alias{wgcna_get_module_genes_by_sign}
\title{Split WGCNA module in module containing only positive or negative correlation}
\usage{
wgcna_get_module_genes_by_sign(wgcna_module, mode)
}
\arguments{
\item{wgcna_module}{Module object that has been produced by \code{wgcna_trait} function}

\item{mode}{Character. "p" or "positive" for positive correlation, "n" or "negative" 
for negative correlation

Only significant colors are used}
}
\value{
Returns a \code{wgcna_module} object with only postively/negatively associated module colors to the trait
}
