% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{wgcna_set_module_size}
\alias{wgcna_set_module_size}
\title{Returns a wgcna module close to \code{size}}
\usage{
wgcna_set_module_size(size, wgcna_module)
}
\arguments{
\item{size}{The}

\item{wgcna_module}{Module object that has been produced by \code{wgcna} function}
}
\description{
Returns a wgcna module close to \code{size}
}
\details{
The function starts with the co-expression module (color) with the lowest 
 p-value and gradually adds more co-expression modules until the size threshold
 has been crossed. Consequently, the resulting module will always be of length at
 least \code{size} 
 
 @return 
 
 \code{wgcna} module object
}
\seealso{
\code{\link{wgcna}}
}
