% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna_trait.R
\name{wgcna_trait}
\alias{wgcna_trait}
\title{An implementation of WGCNA to correlate coexpression modules to disease}
\usage{
wgcna_trait(MODifieR_input, minModuleSize = 30, deepSplit = 2,
  pamRespectsDendro = F, mergeCutHeight = 0.1, numericLabels = T,
  min_KME = 0, pval_cutoff = 0.05, save_toms = T,
  dataset_name = deparse(substitute(MODifieR_input)))
}
\arguments{
\item{MODifieR_input}{An input object created by \link{wgcna} function}
}
\description{
An implementation of WGCNA to correlate coexpression modules to disease
}
\seealso{
\code{\link{create_input}}
}
